Modification de blocs conditionnels par des switch. Choix définitif des types des différents champs  (ID, Fanion). Ces derniers sont des entiers non signés et codés sur 1 octet. Il a été choisi le type short.

+++ PC2RON_Client/src/fr/pc2ron/trame/TrameFactory.java
@@ -51,6 +51,6 @@ public class TrameFactory implements ITrameFactory {
 
 	@Override
 	public ITrame getTrameEnd() {
-		return getTrame(ETypeTrame.TrameFin.getType());
+		return getTrame((byte) ETypeTrame.TrameFin.getType());
 	}
 }
